requirements.txt
# Libraries to install
requests 
pillow

python-dotenv
aiohttp